// frontend/src/lib/api/_generated/orchestrator.d.ts
export interface components {
    schemas: {
// ...
        CurveCreate: {
            /** Name */
            name: string;
            /** Currency */
            currency?: string | null;
            /** Day Counter */
            day_counter?: string | null;
            /** Helper Kind */
            helper_kind?: string | null;
            /** Reference Date */
            reference_date?: string | null;
            /** Points */
            points?: {
                [key: string]: unknown;
            }[];
            /** Body */
            body?: {
                [key: string]: unknown;
            };
        };
// ...
        CurveResponse: {
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
            /**
             * Updated At
             * Format: date-time
             */
            updated_at: string;
            /** Deleted At */
            deleted_at?: string | null;
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /** Name */
            name: string;
            /** Currency */
            currency?: string | null;
            /** Day Counter */
            day_counter?: string | null;
            /** Helper Kind */
            helper_kind?: string | null;
            /** Reference Date */
            reference_date?: string | null;
            /** Points */
            points: {
                [key: string]: unknown;
            }[];
            /** Body */
            body: {
                [key: string]: unknown;
            };
        };
// ...
        CurveUpdate: {
            /** Name */
            name?: string | null;
            /** Currency */
            currency?: string | null;
            /** Day Counter */
            day_counter?: string | null;
            /** Helper Kind */
            helper_kind?: string | null;
            /** Reference Date */
            reference_date?: string | null;
            /** Points */
            points?: {
                [key: string]: unknown;
            }[] | null;
            /** Body */
            body?: {
                [key: string]: unknown;
            } | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}